continued edits to organization and code

+++ GPTPackage/Pegasus.py
@@ -15,48 +15,77 @@ import numpy as np
 import scipy.optimize
 
 
-
-def simulate_gun_to_screen4(mr, cmd_output=True):
-    # Simulate gun to screen4
+def system(call, mr):
+    ''' Execute subprocess call for GPT'''
     write_MRfile('pegasus.mr', mr)
+    cmd = subprocess.run(call.split(' '), capture_output=True, encoding='UTF-8')
+    print(cmd.stderr)
+    
+    return
+
+    
+
+def simulate_gun_to_screen4(mr):
+    # Simulate gun to screen4
     gpt_call = 'mr -o gun_to_screen4.gdf pegasus.mr gpt gun_to_screen4.in'
-    cmd = subprocess.run(gpt_call.split(' '), capture_output=True, encoding='UTF-8')
-    if cmd_output:
-        print(cmd.stderr)
+    system(gpt_call, mr)
     
     # Create gdf for screen at screen4
-    consts, params, particle = load_gdf('gun_to_screen4.gdf')
-    particle[particle.t.notnull()].to_csv('temp.txt', sep=' ', index=False)
-    cmd = subprocess.run('asci2gdf -o beam_at_screen4.gdf temp.txt'.split(' '), capture_output=True, encoding='UTF-8')
-    if cmd_output:
-        print(cmd.stderr)
-    os.remove('temp.txt')
+    write_screens_to_gdf('gun_to_screen4.gdf', 'beam_at_screen4.gdf')
 
     # Update quads
     if mr.estimate_quads:
-        G = particle.G[particle.t.notnull()].mean()
-        s_beam = particle[['x','Bx','y','By']][particle.t.notnull()].cov().values
+        _, _, particle = load_gdf('beam_at_screen4.gdf')
+        G = particle.G.mean()
+        s_beam = particle[['x','Bx','y','By']].cov().values
         mr.quads = estimate_quads(s_beam, G, 5.16-3.191)
     return 
 
 
-def simulate_screen4_to_und(mr, cmd_output=True):
+def simulate_screen4_to_und(mr):
     # Simulate screen4 to und
-    write_MRfile('pegasus.mr', mr)
-    gpt_call = 'mr -o screen4_to_und.gdf pegasus.mr gpt screen4_to_und.in'
-    cmd = subprocess.run(gpt_call.split(' '), capture_output=True, encoding='UTF-8')
-    if cmd_output:
-        print(cmd.stderr)
+    _, _, particle = load_gdf('beam_at_screen4.gdf')
+    gpt_call = f'mr -o screen4_to_und.gdf pegasus.mr gpt screen4_to_und.in tstart={particle.t.mean()}'
+    system(gpt_call, mr)
     
     # Create gdf for screen at und entrance
-    consts, params, particle = load_gdf('gun_to_screen4.gdf')
-    particle[particle.t.notnull()].to_csv('temp.txt', sep=' ', index=False)
-    cmd = subprocess.run('asci2gdf -o beam_at_und.gdf temp.txt'.split(' '), capture_output=True, encoding='UTF-8')
-    if cmd_output:
-        print(cmd.stderr)
-    os.remove('temp.txt')
+    write_screens_to_gdf('screen4_to_und.gdf', 'beam_at_und.gdf')
     return 
 
+
+def simulate_und_to_end(mr, passnum=1):
+    # Simulate und to end
+    _, _, particle = load_gdf('beam_at_und.gdf')
+    gpt_call = f'mr -o und_to_end.gdf pegasus.mr gpt und_to_end.in tstart={particle.t.mean()} passnum={passnum}'
+    system(gpt_call, mr)
+    
+    # Create gdf for screen after und entrance
+    write_screens_to_gdf('und_to_end.gdf', 'beam_after_und.gdf')
+    return 
+
+
+def write_screens_to_gdf(gdf_file, dest_files):
+    ''' Find screen outputs in gdf file and save each to locations specified
+    by dest_files. Length of dest_files must match number of screen outputs.'''
+    # Ensure list type
+    if not isinstance(dest_files, list):
+        dest_files = [dest_files]
+    
+    # Loop over screen outputs
+    _, _, particle = load_gdf(gdf_file)
+    
+    if 't' not in particle.columns:
+        print(f'No screen outputs found in {gdf_file}')
+        return
+    
+    for j, p_idx in enumerate(particle.t.dropna().unstack().index):
+        particle.loc[p_idx].to_csv('temp.txt', sep=' ', index=False)
+        cmd = subprocess.run(f'asci2gdf -o {dest_files[j]} temp.txt'.split(' '), capture_output=True, encoding='UTF-8')
+        print(cmd.stderr)
+        os.remove('temp.txt')
+        
+    return
+    
 
 def write_MRfile(file, mr):
     ''' Writes the numerical data in mr to the file specified.'''
